requirements.txt
numpy
scikit-learn
scikit-image
pillow
tensorflow-gpu
keras
pynput
h5py
